Move common functions into perf_comparison.py, YOLO code into opencv_openvino.py
